backend/requirements.txt
flask
flask-socketio
python-dotenv
openai
eventlet
requests
openai-agents